#000 | REFACTOR | add gateway verticle
